Deleted Camera project folder, added camera functionality to main project
