--- Camera/.vs/ProjectSettings.json
@@ -1,3 +1,0 @@
-{
-  "CurrentProjectSetting": null
-}
