--- Camera/app/src/main/res/drawable/circle.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:innerRadius="0dp"
-    android:shape="ring"
-    android:thicknessRatio="1.9"
-    android:useLevel="false" >
-    <solid android:color="@android:color/transparent" />
-    <stroke
-        android:width="10dp"
-        android:color="@android:color/holo_red_dark" />
-</shape>
