--- Camera/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">Camera</string>
-    <string name="capture_image">Capture Image</string>
-</resources>
