+++ D4App/app/src/main/AndroidManifest.xml
@@ -3,6 +3,8 @@
     package="com.example.d4app">
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.CAMERA"/>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
 
     <application
         android:allowBackup="true"
@@ -14,14 +16,14 @@
         <activity android:name=".DisplayUserInfoActivity">
         </activity>
         <activity android:name=".FormFields">
-
-        </activity>
-        <activity android:name=".DisplayMessageActivity" />
-        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".DisplayMessageActivity" />
+        <activity android:name=".MainActivity">
+
         </activity>
     </application>
