+++ D4App/app/src/main/java/com/example/d4app/FormFields.kt
@@ -1,18 +1,25 @@
 package com.example.d4app
 
+import android.Manifest
+import android.app.Activity
 import android.app.DatePickerDialog
 import android.app.DatePickerDialog.OnDateSetListener
+import android.content.ContentValues
 import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.icu.util.Calendar
+import android.net.Uri
 import android.os.Build
 import android.os.Bundle
+import android.provider.MediaStore
 import android.text.InputType
 import android.view.View
 import android.widget.*
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.*
+import kotlinx.android.synthetic.main.activity_form_fields.*
 import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import java.io.IOException
@@ -20,10 +27,12 @@ import java.io.IOException
 
 class FormFields : AppCompatActivity() {
 
+    private val PERMISSION_CODE = 1000;
+    private val IMAGE_CAPTURE_CODE = 1001
+    var image_uri: Uri? = null
+
     var picker: DatePickerDialog? = null
     var birthDay: EditText? = null
-    var btnGet: Button? = null
-    var tvw: TextView? = null
 
     var userAge: Int? = null
 
@@ -35,7 +44,6 @@ class FormFields : AppCompatActivity() {
         // SETUP ELEMENTS
         val spinner: Spinner = findViewById(R.id.userSex)
 
-        tvw = findViewById<TextView>(R.id.textView1)
         birthDay = findViewById<EditText>(R.id.userBDate)
 
         val name = findViewById<EditText>(R.id.userName)
@@ -44,6 +52,31 @@ class FormFields : AppCompatActivity() {
         val weight = findViewById<EditText>(R.id.userWeight)
         val bioRole = findViewById<Spinner>(R.id.userSex)
         val btnSave = findViewById<Button>(R.id.btnSaveUserProfile)
+
+
+        // Camera functionality
+        capture_btn.setOnClickListener {
+            //if system os is Marshmallow or Above, we need to request runtime permission
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M){
+                if (checkSelfPermission(Manifest.permission.CAMERA)
+                    == PackageManager.PERMISSION_DENIED ||
+                    checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                    == PackageManager.PERMISSION_DENIED){
+                    //permission was not enabled
+                    val permission = arrayOf(Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                    //show popup to request permission
+                    requestPermissions(permission, PERMISSION_CODE)
+                }
+                else{
+                    //permission already granted
+                    openCamera()
+                }
+            }
+            else{
+
+                openCamera()
+            }
+        }
 
 
         // Create an ArrayAdapter using the string array and a default spinner layout
@@ -79,20 +112,17 @@ class FormFields : AppCompatActivity() {
             picker!!.show()
         }
 
-        btnGet = findViewById<View>(R.id.button1) as Button
-        btnGet!!.setOnClickListener { tvw!!.text = "Selected Date: " + birthDay!!.text }
-
 
         btnSave.setOnClickListener(View.OnClickListener {
-        // USED for debugging  Toast.makeText(applicationContext,"data save:" + bioRole.selectedItem.toString(),Toast.LENGTH_LONG).show()
+            // USED for debugging  Toast.makeText(applicationContext,"data save:" + bioRole.selectedItem.toString(),Toast.LENGTH_LONG).show()
 
 
             val userData: String =
                 "name:" + name.text.toString() + ",birthdate:" + birthDay.toString() + "age:" + userAge + ",height:" + heightFeet.text.toString() + "-" + heightInches.text.toString() + ",weight:" + weight.text.toString() + ",biorole:" + bioRole.selectedItem.toString()
-             writeDataToFile(userData)
+            writeDataToFile(userData)
 
-           val intent = Intent(this, DisplayUserInfoActivity::class.java)
-           startActivity(intent)
+            val intent = Intent(this, DisplayUserInfoActivity::class.java)
+            startActivity(intent)
         })
     }
 
@@ -129,6 +159,47 @@ class FormFields : AppCompatActivity() {
         }
         val ageInt = age
         return ageInt
+    }
+
+
+    // camera functionality
+
+    private fun openCamera() {
+        val values = ContentValues()
+        values.put(MediaStore.Images.Media.TITLE, "New Picture")
+        values.put(MediaStore.Images.Media.DESCRIPTION, "From the Camera")
+        image_uri = contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values)
+        //camera intent
+        val cameraIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+        cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, image_uri)
+        startActivityForResult(cameraIntent, IMAGE_CAPTURE_CODE)
+    }
+
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
+        //called when user presses ALLOW or DENY from Permission Request Popup
+        when(requestCode){
+            PERMISSION_CODE -> {
+                if (grantResults.size > 0 && grantResults[0] ==
+                    PackageManager.PERMISSION_GRANTED){
+                    //permission from popup was granted
+                    openCamera()
+                }
+                else{
+                    //permission from popup was denied
+                    Toast.makeText(this, "Permission denied", Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        //called when image was captured from camera intent
+        if (resultCode == Activity.RESULT_OK){
+            //set image captured to image view
+            image_view.setImageURI(image_uri)
+        }
     }
